requirements.txt
# Core LLM backends
anthropic>=0.25.0
openai>=1.0.0
google-genai>=0.5.0

# Vector search & embeddings (no LangChain)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0

# Document ingestion
PyPDF2>=3.0.0
python-docx>=0.8.11

# Export
weasyprint>=53.0
jinja2>=3.0

# Utilities
requests>=2.28.0
backoff>=1.10.0
pyyaml>=6.0
